+++ R/lb-keogh.R
@@ -61,7 +61,7 @@ lb_keogh <- function(x, y, window.size = NULL, norm = "L1",
         if (error.check) check_consistency(y, "ts")
         window.size <- check_consistency(window.size, "window")
 
-        envelopes <- compute_envelop(y, window.size = window.size, error.check = FALSE)
+        envelopes <- compute_envelope(y, window.size = window.size, error.check = FALSE)
         lower.env <- envelopes$lower
         upper.env <- envelopes$upper
 
@@ -111,7 +111,7 @@ lb_keogh_proxy <- function(x, y = NULL, window.size = NULL, norm = "L1", ...,
     if (is_multivariate(x) || is_multivariate(y))
         stop("lb_keogh does not support multivariate series.")
 
-    envelopes <- lapply(y, function(s) { compute_envelop(s, window.size, error.check = FALSE) })
+    envelopes <- lapply(y, function(s) { compute_envelope(s, window.size, error.check = FALSE) })
     lower.env <- lapply(envelopes, "[[", "lower")
     upper.env <- lapply(envelopes, "[[", "upper")
     lower.env <- split_parallel(lower.env)